+++ src/lib/adsb/CMakeLists.txt
@@ -39,7 +39,7 @@ set(adsb_srcs
 if(CONFIG_NAVIGATOR_ADSB_F3442)
 	list(APPEND adsb_srcs f3442_standard/DaaF3442.cpp)
 else()
-	list(APPEND adsb_srcs crosstrack_based_daa/DaaCrosstrack.cpp)
+	list(APPEND adsb_srcs crosstrack_standard/DaaCrosstrack.cpp)
 endif()
 
 px4_add_library(adsb ${adsb_srcs})
@@ -54,5 +54,5 @@ px4_add_unit_gtest(SRC DaaEncodedIdTest.cpp LINKLIBS adsb)
 if(CONFIG_NAVIGATOR_ADSB_F3442)
 	px4_add_functional_gtest(SRC f3442_standard/DaaF3442Test.cpp LINKLIBS adsb)
 else()
-	px4_add_functional_gtest(SRC crosstrack_based_daa/DaaCrosstrackTest.cpp LINKLIBS adsb)
+	px4_add_functional_gtest(SRC crosstrack_standard/DaaCrosstrackTest.cpp LINKLIBS adsb)
 endif()
